+++ src/querier/cursors.zig
@@ -1,7 +1,7 @@
 const std = @import("std");
 const sqlite = @import("../sqlite.zig");
 
-pub fn getLastPollerHeight(conn: *sqlite.Conn) !u64 {
+pub fn getLastPollerHeight(conn: *sqlite.Conn) !u32 {
     const result = conn.row("SELECT block_number FROM cursors WHERE id = 1;", .{}) catch |err| {
         std.log.err("failed to fetch latest poller height: {}", .{err});
         return err;
@@ -15,6 +15,6 @@ pub fn getLastPollerHeight(conn: *sqlite.Conn) !u64 {
     return 0;
 }
 
-pub fn upsertPollerHeight(conn: *sqlite.Conn, new_height: u64) !void {
+pub fn upsertPollerHeight(conn: *sqlite.Conn, new_height: u32) !void {
     try conn.exec("REPLACE INTO cursors(id, block_number) VALUES(1, ?1);", .{new_height});
 }
