+++ src/querier/stakers.zig
@@ -4,7 +4,7 @@ const ArenaAllocator = std.heap.ArenaAllocator;
 const sqlite = @import("../sqlite.zig");
 const Envelope = @import("common.zig").Envelope;
 
-pub fn insertNewStaker(conn: *sqlite.Conn, address: []const u8, epoch_number: u64, stake_balance: u64, stake_percentage: f64) !void {
+pub fn insertNewStaker(conn: *sqlite.Conn, address: []const u8, epoch_number: u32, stake_balance: u64, stake_percentage: f64) !void {
     try conn.exec("REPLACE INTO stakers(address, epoch_number, stake_balance, stake_percentage) VALUES(?1, ?2, ?3, ?4);", .{ address, epoch_number, stake_balance, stake_percentage });
 }
 
@@ -13,7 +13,7 @@ pub const GetStakersByEpochRow = struct {
     stake_percentage: f64,
 };
 
-pub fn getStakersByEpoch(conn: *sqlite.Conn, allocator: Allocator, epoch_number: u64) !Envelope(GetStakersByEpochRow) {
+pub fn getStakersByEpoch(conn: *sqlite.Conn, allocator: Allocator, epoch_number: u32) !Envelope(GetStakersByEpochRow) {
     var rows = try conn.rows("SELECT address, stake_percentage FROM stakers WHERE epoch_number = ?1;", .{epoch_number});
     defer rows.deinit();
